refactor: change spent amount from integer to bigint

+++ db/migrate/20240102033054_create_spents.rb
@@ -1,10 +1,11 @@
 class CreateSpents < ActiveRecord::Migration[7.1]
   def change
     create_table :spents do |t|
-      t.integer :amount
+      t.bigint :amount
       t.string :name
       t.references :category
       t.references :subcategory
+
       t.timestamps
     end
   end
